services/syndy/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
boto3==1.34.9
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
s3fs